+++ src/actions/conversation.js
@@ -562,17 +562,3 @@ export const unAssignConversation =
       dispatch({ type: ASSIGN_CONVERSATION_ERROR });
     }
   };
-export const getAllCustomAttributes = () => async dispatch => {
-  dispatch({ type: GET_ALL_CUSTOM_ATTRIBUTES });
-  try {
-    const apiUrl = 'custom_attribute_definitions';
-    const response = await axios.get(apiUrl);
-    const { data } = response;
-    dispatch({
-      type: GET_ALL_CUSTOM_ATTRIBUTES_SUCCESS,
-      payload: data,
-    });
-  } catch (error) {
-    dispatch({ type: GET_ALL_CUSTOM_ATTRIBUTES_ERROR, payload: error });
-  }
-};
